refactor: consolidate entity resolution into single handler file
- Merge mock_entity_resolution.py service into entity_resolution.py handler
- Rename internal functions with underscore prefix for clarity
- Maintain same functionality, just single file for consistency
- Aligns with existing codebase pattern (schema.py handler is 1100+ lines)
